Rewrite of channel
Add live support, updated README and reorganized channel structure to
make maintanence easier
